feat: Hand `globalThis` to the core configuration
It seems that it's useful for some core extensions to have controlled access to
the global scope, although it shouldn't need to know what specific environment
it runs in, which is the prerogative of the base library.

A good example is providing an adapter for `document.querySelector` for your
widgets.